Change Windows support to VS 2019
* VS 2022 has a bug where openssl/1.1.0l recipe will invoke the incorrect nmake binary when targeting x86_64 (it invokes the x86 binary)

+++ src/BuildSystem/build.ps1
@@ -4,32 +4,45 @@ param
     $BuildType = "Debug"
 )
 
-# enable echoing
-Set-PSDebug -Trace 1
+$ProjectPath = "$PSScriptRoot/../.."
 
-$BuildTypeLower = $BuildType.ToLower()
-$ConanConfigPath = "conan-config"
-$ConanProfilePath = "$ConanConfigPath/profiles/ergonomic-cpp-$BuildTypeLower"
-# FIXME Windows
-#$BuildPath = "build"
-$BuildPath = "build/$BuildTypeLower"
-$CmakeBuildType = (Get-Culture).TextInfo.ToTitleCase($BuildTypeLower)
-
-# move to repo root
-Set-Location "$PSScriptRoot/../.."
-
-# installing profiles isn't necessary, but helpful for setting up CLion
-Remove-Item -Force -Path "$ConanConfigPath/.DS_Store" -ErrorAction Ignore
+# installing profiles is a dependency for installing packages since the profiles have
+# shared settings that affect which package versions are downloaded
+if ($IsWindows)
+{
+    $ConanConfigPath = "$ProjectPath/conan-config/Windows"
+}
+else
+{
+    $ConanConfigPath = "$ProjectPath/conan-config/Default"
+    Remove-Item -Force -Path "$ConanConfigPath/.DS_Store" -ErrorAction Ignore
+}
 conan config install $ConanConfigPath
 
-# create and cd to the BuildPath
+# create and cd to the BuildPath since we need to run conan install and cmake there
+$BuildTypeLower = $BuildType.ToLower()
+$BuildType = (Get-Culture).TextInfo.ToTitleCase($BuildTypeLower)
+if ($IsWindows)
+{
+    $BuildPath = "$ProjectPath/build"
+}
+else
+{
+    $BuildPath = "$ProjectPath/build/$BuildTypeLower"
+}
 New-Item -Type Directory -Force $BuildPath
 Set-Location $BuildPath
-# it's easier to install from the BuildPath since we need to run cmake here
-conan install ../.. -pr="../../$ConanProfilePath" -pr="default" --build=missing
-# the source dir is the parent since that's where CMakeLists.txt is
-cmake -DCMAKE_BUILD_TYPE=$CmakeBuildType -S ../.. -B . -G "Unix Makefiles"
-make all
-# FIXME Windows
-#cmake -S ../.. -B . -G "Visual Studio 17 2022"
-#msbuild SimpleCmake.sln /p:Configuration=$CmakeBuildType
+
+conan install $ProjectPath -pr="ergonomic-cpp-$BuildTypeLower" --build=missing
+
+$env:CMAKE_BUILD_TYPE = $BuildType
+if ($IsWindows)
+{
+    cmake -S $ProjectPath -B . -G "Visual Studio 16 2019" -A x64
+    msbuild ErgonomicCpp.sln /p:Configuration=$BuildType
+}
+else
+{
+    cmake -D -S $ProjectPath -B . -G "Unix Makefiles"
+    make all
+}
